Replace proprietary dxil.dll code signing blob with FOSS alternative
Normal usage of dxcompiler requires that you use a proprietary `dxil.dll`
distributed by Microsoft. If this is not found next to the dxcompiler binary
when compiling shaders, then the final shader binary will not be signed and
as a result will not run on others' machines.

DXC is also written with the assumption that dxil.dll and dxcompiler.dll itself
are loaded as dynamic libraries. This, combined with dxil.dll being a proprietary
blob, is annoying for users of Mach engine and [many others](https://bevy-cheatbook.github.io/platforms/windows.html#dxc-compiler-support)
as it just complicates application distribution further.

This commit replaces dxcompiler.dll runtime loading behavior with an 'emulated'
dynamic library, effectively completing our effort to make dxcompiler a static
library.

dxil.dll is closed-source, so we cannot simply patch it in the same way. To fix
this, we outright disable runtime loading of dxil.dll and silence warnings related
to it not being present / the final binary not being code signed. Instead, once
the compiler would emit a compiled shader blob, we perform our own code signing
algorithm (Mach Siegbert Vogt DXCSA) which results in a bitwise identical compiled
shader, and thus dxil.dll is no longer needed to perform code signing of shaders.

Ultimately, this means mach-dxcompiler can be a fully static library, code signing
included.

Signed-off-by: Stephen Gutekanst <stephen@hexops.com>